Make `demo.sh` executable
This changes the permissions of `demo.sh` to be an executable
script (using `$ sudo chmod +x demo.sh`).

Additionally, `zombie.out` is used inside the script as part of
the demo, but is not present in the repo. This adds a check to
have `zombie.c` compiled into `zombie.out`, which is finally
deleted just before the script terminates.

Note: The file is compiled without any compiler flags.
`$ gcc zombie.c -o zombie.out`