Split off color format macros into `printk_colors.h`
The color formatting macros used to colorize the output of `printk()`
were not strictly relevant to the actual task listing functionality,
and cluttered the source file. Instead, this splits them off into
a dedicated header file `printk_colors.h`, which we simply include
to get access to the macros.

Nothing else changes, including the existing functionality.