Add function to display tasks in depth-first search (DFS) order
The original functionality printed out the tasks in linear order
(that is, in ascending order of their PIDs). However, tasks could
also be iterated through using the process tree, which holds processes
and their children.

We use the `list_for_each()` macro to access each child of the
current task. We can iterate over the entire tree by recursively
invoking our function. We can use the provided `init_list` to
initially invoke the function on the init process.