Refactor task printing lines into separate functions
The functions for iterating through and printing tasks used to
directly store and handle the logic for formatting the output,
such as the header row and the format for each task. These have
now been split off into 2 dedicated functions:
- `print_task_row_header()`
- `print_task_details(struct task_struct *task)`

The first prints the header row that displays the columns to be
populated by the tasks' details.

The second takes a task as input, and prints out its details in
a formatted manner.

The rest of the functionality remains the same. The linear and DFS
methods now invoke the above 2 functions instead of directly doing
this themselves.