Configure Makefile to choose task list mode when compiling
Earlier, the task listing mode was manually specified by the choice
of function invoked inside the module init function.

This commit has the choice of listing mode made when building the
module from source. The default mode chosen is linear (or if the
macro `TASK_LIST_MODE_LINEAR` is defined). The mode used can be
set to DFS by defining the macro `TASK_LIST_MODE_DFS`.

`Makefile` and `demo.sh` are also modified to take this into account
and configure the build accordingly.