Minor change to `get_task_state_name()`
The flag `TASK_RUNNING` is already defined in `<linux/sched.h>`;
there is no need for us to explicitly define it inside the logic
of `get_task_state_name()`.

Minor note: the macro `list_for_each()` is defined in the header
`<linux/list.h>`. However, the module runs normally without needing
to explicitly #include it. Additionally, it uses `struct list_head`,
which is itself defined elsewhere.
There is no need to add anything to the include list at this time.