Specify conditional compilation for module, and usage for demo script
This more clearly specifies the conditional compilation for the
kernel module. Now, unless the macro `TASK_LIST_MODE_DFS` is
defined at compilation, the kernel module will use the linear
printing order. Only when the macro is defined will it switch
to using the DFS printing order.

Also, the usage for `demo.sh` is specified in the script itself.
Now, the demo has to be explicitly called with the single argument
of either `linear` or `dfs`. These usage instructions will be
printed when invoking the script in any other way.